requirements.txt
Django==1.10.3

# db
psycopg2==2.6.2

# api
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
